requirements.txt
selenium==4.11.2
pytest==7.4.0
chromedriver-autoinstaller==0.4.0
wait-on
websocket-client==1.5.1